Reogranisation
The files and folders are being shifted around because I'm going to include Visual Studio solution and project files. The headers have been moved into a separate folder because I will now be taking advantage of VS's systems (whereas before I was trying to make it VS independant which, as it turns out, made it a pain to compile for others).